Module google_clouderrorreporting1_beta1::api

Structs

Clouderrorreporting
Central instance to access all Clouderrorreporting related resource activities

DeleteEventsResponse
Response message for deleting error events.

ErrorContext
A description of the context in which an error occurred. This data should be provided by the application when reporting an error, unless the error report has been generated automatically from Google App Engine logs.

ErrorEvent
An error event which is returned by the Error Reporting system.

ErrorGroup
Description of a group of similar error events.

ErrorGroupStats
Data extracted for a specific group based on certain filter criteria, such as a given time period and/or service filter.

HttpRequestContext
HTTP request data that is related to a reported error. This data should be provided by the application when reporting an error, unless the error report has been generated automatically from Google App Engine logs.

ListEventsResponse
Contains a set of requested error events.

ListGroupStatsResponse
Contains a set of requested error group stats.

ProjectDeleteEventCall
Deletes all error events of a given project.

ProjectEventListCall
Lists the specified events.

ProjectEventReportCall
Report an individual error event and record the event to a log. This endpoint accepts either an OAuth token, or an API key for authentication. To use an API key, append it to the URL as the value of a key parameter. For example: POST https://clouderrorreporting.googleapis.com/v1beta1/{projectName}/events:report?key=123ABC456 Note: Error Reporting is a global service built on Cloud Logging and doesn't analyze logs stored in regional log buckets or logs routed to other Google Cloud projects. For more information, see Using Error Reporting with regionalized logs.

ProjectGroupGetCall
Get the specified group.

ProjectGroupStatListCall
Lists the specified groups.

ProjectGroupUpdateCall
Replace the data for the specified group. Fails if the group does not exist.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Clouderrorreporting hub.

ReportErrorEventResponse
Response for reporting an individual error event. Data may be added to this message in the future.

ReportedErrorEvent
An error event which is reported to the Error Reporting system.

ServiceContext
Describes a running service that sends errors. Its version changes over time and multiple versions can run in parallel.

SourceLocation
Indicates a location in the source code of the service for which errors are reported. functionName must be provided by the application when reporting an error, unless the error report contains a message with a supported exception stack trace. All fields are optional for the later case.

SourceReference
A reference to a particular snapshot of the source tree used to build and deploy an application.

TimedCount
The number of errors in a given time period. All numbers are approximate since the error events are sampled before counting them.

TrackingIssue
Information related to tracking the progress on resolving the error.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.